requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0

# Data and visualization
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Medical imaging
nibabel>=5.1.0
pydicom>=2.4.0
SimpleITK>=2.3.0

# Metrics and logging
tensorboard>=2.13.0
tqdm>=4.65.0
pyyaml>=6.0

# Utilities
timm>=0.9.0
albumentations>=1.3.0
grad-cam>=1.4.8
